requirements.txt
tensorboard==2.1.1
numpy==1.18.2
Pillow==7.1.1
Keras==2.3.1
requests==2.23.0
Python==3.7.7rc1
